requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
